feat: add ImportedSheet, multi-select export, rename ExportSolid->ExportSelected
- Add wy3d::ImportedSheet class (parallels ImportedSolid for sheet bodies)
- Extend Importer to create both ImportedSolid and ImportedSheet
- Support multi-selection in ExportSelectedCommand
- Rename ExportSolidCommand to ExportSelectedCommand
- Simplify Exporter API to perform(Database*) + perform(TopoDS_Shape&)
- Add Import/Export/Imported Sheet translations
- Register ImportedSheet in snap, gizmo, feature tree

+++ src/wy3dApp/exporter/Importer.cpp
@@ -21,6 +21,7 @@
 #include <wydbDatabase.h>
 #include <wydbTransaction.h>
 #include <wy3dImportedSolid.h>
+#include <wy3dImportedSheet.h>
 #include "application/Application.h"
 #include "scene/Scene.h"
 
@@ -53,17 +54,21 @@ bool Importer::perform(wydb::Database* pDb, const std::wstring& fileFullPath)
         assert(false);
         goto ABORT_TRANS;
     }
-    pDb->getTransactionManager()->endTransaction();
 
+    wy3d::ImportedSheet* pImportedSheet(nullptr);
+    if (wy::ErrorStatus::Ok != wy3d::ImportedSheet::create(pTrans, fileFullPath, pImportedSheet) || !pImportedSheet)
+    {
+        assert(false);
+        goto ABORT_TRANS;
+    }
+
+    pDb->getTransactionManager()->endTransaction();
     return true;
 
 ABORT_TRANS:
     assert(false);
     pDb->getTransactionManager()->abortTransaction();
-    if (pImportedSolid)
-    {
-        wydb::deleteElement(pImportedSolid);
-        pImportedSolid = nullptr;
-    }
+    if (pImportedSolid) { wydb::deleteElement(pImportedSolid); pImportedSolid = nullptr; }
+    if (pImportedSheet) { wydb::deleteElement(pImportedSheet); pImportedSheet = nullptr; }
     return false;
 }
